docs: CustomerService, ServiceImpl 분리
