+++ src/main/java/org/boot/growup/common/SecurityConfig.java
@@ -51,11 +51,6 @@ public class SecurityConfig {
                 .authorizeHttpRequests(auth -> auth
                         .requestMatchers(requestMatcherHolder.getRequestMatchersByMinRole(null)).permitAll()
                         .requestMatchers(requestMatcherHolder.getRequestMatchersByMinRole(Role.CUSTOMER)).hasRole("CUSTOMER")
-//                        .requestMatchers("/customers/email/**", "/customers/oauth/**").permitAll()
-//                        .requestMatchers("/sellers/**").permitAll()
-//                        .requestMatchers("/v3/**", "/swagger-ui/**").permitAll() // swagger 설정
-//                        .requestMatchers("/login/**", "/images/**", "/favicon.ico").permitAll()
-//                        .requestMatchers("/customers/**").hasRole("CUSTOMER")
                         .anyRequest().authenticated()
                 )
                 .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
